+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - bump duckdb to 1.0.0.
 - add `DuckDB::ExtractedStatements` class.
 - add `DuckDB::ExtractedStatements#size`.
+- add `DuckDB::ExtractedStatements#prepared_statement`.
 - raise error when `DuckDB::ExtractedStatements#new` is called with invalid SQL.
 - The following public/private methods will be deprecated.
   - `DuckDB::Result#streaming?`